+++ nal/src/main/java/nars/table/TemporalBeliefTable.java
@@ -20,9 +20,9 @@ public interface TemporalBeliefTable extends TaskTable, Iterable<Task> {
 
 
     /** finds the strongest match to the specified parameters. Task against is an optional argument which can be used to compare internal temporal dt structure for similarity */
-    @Nullable Task match(long when, long now, float dur, @Nullable Task against);
+    @Nullable Task match(long when, long now, int dur, @Nullable Task against);
 
-    @Nullable Truth truth(long when, long now, float dur, EternalTable eternal);
+    @Nullable Truth truth(long when, long now, int dur, EternalTable eternal);
 
     @Nullable Task add(@NotNull Task input, EternalTable eternal, Concept concept, @NotNull NAR nar);
 
@@ -63,12 +63,12 @@ public interface TemporalBeliefTable extends TaskTable, Iterable<Task> {
         }
 
         @Override
-        public @Nullable Task match(long when, long now, float dur, @Nullable Task against) {
+        public @Nullable Task match(long when, long now, int dur, @Nullable Task against) {
             return null;
         }
 
         @Override
-        public @Nullable Truth truth(long when, long now, float dur, EternalTable eternal) {
+        public @Nullable Truth truth(long when, long now, int dur, EternalTable eternal) {
             return null;
         }
